pyproject.toml
[project]
name = "gc-ope"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Xudong Gong", email = "gongxudong_cs@aliyun.com" },
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flycraft>=0.1.9",
    "gymnasium>=1.2.1",
    "gymnasium-robotics>=1.4.1",
    "hydra-core>=1.3.2",
    "numpy>=2.3.4",
    "omegaconf>=2.3.0",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.3",
    "seaborn>=0.13.2",
    "stable-baselines3>=2.7.0",
    "tensorboard>=2.20.0",
    "torch>=2.9.0",
    "tqdm>=4.67.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gc_ope"]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
]
